refactor: move package of internal interfaces
