+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/Embrace.java
@@ -13,7 +13,14 @@ import java.util.Map;
 import io.embrace.android.embracesdk.annotation.InternalApi;
 import io.embrace.android.embracesdk.internal.EmbraceInternalInterface;
 import io.embrace.android.embracesdk.internal.Systrace;
+import io.embrace.android.embracesdk.internal.api.EmbraceAndroidApi;
+import io.embrace.android.embracesdk.internal.api.EmbraceApi;
+import io.embrace.android.embracesdk.internal.api.LogsApi;
+import io.embrace.android.embracesdk.internal.api.MomentsApi;
+import io.embrace.android.embracesdk.internal.api.NetworkRequestApi;
 import io.embrace.android.embracesdk.internal.api.SdkStateApi;
+import io.embrace.android.embracesdk.internal.api.SessionApi;
+import io.embrace.android.embracesdk.internal.api.UserApi;
 import io.embrace.android.embracesdk.network.EmbraceNetworkRequest;
 import io.embrace.android.embracesdk.spans.EmbraceSpan;
 import io.embrace.android.embracesdk.spans.EmbraceSpanEvent;
