+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/LogType.kt
@@ -1,0 +1,14 @@
+package io.embrace.android.embracesdk
+
+/**
+ * Deprecated: use Severity instead. This enum is deprecated and will be removed in
+ * a future release.
+ *
+ * Will flag the message as one of info, warning, or error for filtering on the dashboard
+ */
+@Deprecated("")
+public enum class LogType {
+    INFO,
+    WARNING,
+    ERROR
+}
