+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/api/OtelExporterApi.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk
+package io.embrace.android.embracesdk.internal.api
 
 import io.opentelemetry.sdk.logs.export.LogRecordExporter
 import io.opentelemetry.sdk.trace.export.SpanExporter
