+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/api/ViewTrackingApi.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk
+package io.embrace.android.embracesdk.internal.api
 
 import android.app.Application
 import io.embrace.android.embracesdk.payload.TapBreadcrumb
