+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/api/delegate/MomentsApiDelegate.kt
@@ -1,9 +1,9 @@
 package io.embrace.android.embracesdk.internal.api.delegate
 
-import io.embrace.android.embracesdk.MomentsApi
 import io.embrace.android.embracesdk.event.EmbraceEventService
 import io.embrace.android.embracesdk.injection.ModuleInitBootstrapper
 import io.embrace.android.embracesdk.injection.embraceImplInject
+import io.embrace.android.embracesdk.internal.api.MomentsApi
 import io.embrace.android.embracesdk.internal.utils.toNonNullMap
 import io.embrace.android.embracesdk.utils.PropertyUtils
