+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/api/delegate/OtelExporterApiDelegate.kt
@@ -1,7 +1,7 @@
 package io.embrace.android.embracesdk.internal.api.delegate
 
-import io.embrace.android.embracesdk.OtelExporterApi
 import io.embrace.android.embracesdk.injection.ModuleInitBootstrapper
+import io.embrace.android.embracesdk.internal.api.OtelExporterApi
 import io.opentelemetry.sdk.logs.export.LogRecordExporter
 import io.opentelemetry.sdk.trace.export.SpanExporter
