+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/api/delegate/UserApiDelegate.kt
@@ -1,8 +1,8 @@
 package io.embrace.android.embracesdk.internal.api.delegate
 
-import io.embrace.android.embracesdk.UserApi
 import io.embrace.android.embracesdk.injection.ModuleInitBootstrapper
 import io.embrace.android.embracesdk.injection.embraceImplInject
+import io.embrace.android.embracesdk.internal.api.UserApi
 
 internal class UserApiDelegate(
     bootstrapper: ModuleInitBootstrapper,
