+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/api/delegate/ViewTrackingApiDelegate.kt
@@ -2,9 +2,9 @@ package io.embrace.android.embracesdk.internal.api.delegate
 
 import android.app.Application
 import io.embrace.android.embracesdk.Embrace
-import io.embrace.android.embracesdk.ViewTrackingApi
 import io.embrace.android.embracesdk.injection.ModuleInitBootstrapper
 import io.embrace.android.embracesdk.injection.embraceImplInject
+import io.embrace.android.embracesdk.internal.api.ViewTrackingApi
 import io.embrace.android.embracesdk.payload.TapBreadcrumb
 
 internal class ViewTrackingApiDelegate(
