+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/InternalInterfaceModuleImplTest.kt
@@ -6,6 +6,8 @@ import io.embrace.android.embracesdk.fakes.injection.FakeCustomerLogModule
 import io.embrace.android.embracesdk.fakes.injection.FakeDataContainerModule
 import io.embrace.android.embracesdk.fakes.injection.FakeEssentialServiceModule
 import io.embrace.android.embracesdk.fakes.injection.FakeInitModule
+import io.embrace.android.embracesdk.injection.InternalInterfaceModule
+import io.embrace.android.embracesdk.injection.InternalInterfaceModuleImpl
 import org.junit.Assert.assertNotNull
 import org.junit.Test
